Optimizations:     * rewrote __getattr__ and __setattr__ to avoid wanton __getattr__ recursion     * added nodeType and nodeName attributes to Node instances, so n.nodeType       references no longer call __getattr__ at all.     * instead of assigning to self.attributes, assign directly to __dict__,       avoiding a __setattr_ call
The net effect of these changes is that the demo/dom/benchmark.py runs
without calling Node's __getattr__ or __setattr__ at all.  This
improves speed almost fourfold; the 212K XML file of "A Winter's Tale"
is parsed in 25 seconds, instead of 79 seconds.

Removed assert statements, replacing them with if statements that
    raise exceptions.