Fixed query-replace error that kept current sysid from being updated when new entities were opened.
The system identifiers of entities are now absolutized when the entity
declarations are read, rather than when they are referred to. This also
means that they are absolutized relative to the parameter entity
in which the declarations appear, rather than relative to the entity from
which they are referred to. This fixes a bug that has always been in the
xmlproc source code.