Support pickling of the ElementInfo objects this module produces as well.
Lots of cleanup, avoiding use of temporary data structures.  This
means that the order of calls to the node filter for DOCTYPE-related
nodes has changed; the DocumentType node is now filtered before the
Entity and Notation nodes, which makes more sense.  It also means that
if the DocumentType node is rejected, the Entity and Notation nodes
are never created or passed to the filter.  These are now handled in
document order, which is consistent with expectations.